Remove unused modules from model/material
