+++ message_ix_models/model/material/build.py
@@ -10,7 +10,7 @@ from message_ix_models.model.material.data_aluminum import gen_data_aluminum
 from message_ix_models.model.material.data_ammonia_new import gen_all_NH3_fert
 from message_ix_models.model.material.data_cement import gen_data_cement
 from message_ix_models.model.material.data_generic import gen_data_generic
-from message_ix_models.model.material.data_methanol_new import gen_data_methanol_new
+from message_ix_models.model.material.data_methanol import gen_data_methanol
 from message_ix_models.model.material.data_petro import gen_data_petro_chemicals
 from message_ix_models.model.material.data_power_sector import gen_data_power_sector
 from message_ix_models.model.material.data_steel import gen_data_steel
@@ -46,8 +46,7 @@ from message_ix_models.util.scenarioinfo import ScenarioInfo, Spec
 log = logging.getLogger(__name__)
 
 DATA_FUNCTIONS = [
-    # gen_data_buildings,
-    gen_data_methanol_new,
+    gen_data_methanol,
     gen_all_NH3_fert,
     gen_data_generic,
     gen_data_steel,
